+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Watch::refresh()` to specify `REFRESH` clause.
 
+### Fixed
+- `derive(Row)`: handle raw identifiers.
+
 ## [0.7.1] - 2021-06-29
 ### Fixed
 - Get rid of "socket is not connected" errors (found by [@let4be](https://github.com/let4be)).